requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coverage==4.5.2
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-countries==5.3.3
django-nose==1.4.6
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==1.0.2
google-api-core==1.10.0
google-auth==1.6.3
google-cloud-bigquery==1.11.2
google-cloud-core==0.29.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
gunicorn==19.9.0
idna==2.8
nose==1.3.7
protobuf==3.7.1
psycopg2==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.7.5
pytz==2018.7
rednose==1.3.0
requests==2.21.0
rsa==4.0
schedule==0.6.0
six==1.12.0
termstyle==0.1.11
text-unidecode==1.2
urllib3==1.24.1
whitenoise==4.1.2